8349351: Combine Screen Inset Tests into a Single File
Backport-of: 9ea81d90175c11460d0efa83f82ceccc4ee2cd3b
